Move Image class to its own source and header files
- Move the Image class to its own source and header files (Image.cpp and Image.h), allowing image processors to depend on Image without depending on ImageHandler and ImageLoader.
- Add a ConstImagePtr alias to clarify the behavior of image processing interfaces.
- Minor whitespace fixes in Util.h.
