+++ source/MaterialXRender/StbImageLoader.h
@@ -42,7 +42,7 @@ class StbImageLoader : public ImageLoader
 
     /// Save an image to the file system.
     bool saveImage(const FilePath& filePath,
-                   ImagePtr image,
+                   ConstImagePtr image,
                    bool verticalFlip = false) override;
 
     /// Load an image from the file system.
