+++ source/MaterialXRender/Util.h
@@ -13,87 +13,87 @@
 #include <MaterialXGenShader/ShaderGenerator.h>
 #include <MaterialXGenShader/Util.h>
 
-#include <MaterialXCore/Document.h>
-
 #include <map>
 
 namespace MaterialX
 {
-    /// @name Shader utilities
-    /// @{
 
-    /// Create a shader for a given element
-    ShaderPtr createShader(const string& shaderName, GenContext& context, ElementPtr elem);
+/// @name Shader Utilities
+/// @{
 
-    /// Create a shader with a constant color output for a given element
-    ShaderPtr createConstantShader(GenContext& context,
-                                   DocumentPtr stdLib,
-                                   const string& shaderName,
-                                   const Color3& color);
+/// Create a shader for a given element.
+ShaderPtr createShader(const string& shaderName, GenContext& context, ElementPtr elem);
 
-    /// @}
-    /// @name User interface utilities
-    /// @{
+/// Create a shader with a constant color output, using the given standard libraries
+/// for code generation.
+ShaderPtr createConstantShader(GenContext& context,
+                               DocumentPtr stdLib,
+                               const string& shaderName,
+                               const Color3& color);
 
-    /// Set of possible UI properties for an element
-    struct UIProperties
-    {
-        /// UI name
-        string uiName;
+/// @}
+/// @name User Interface Utilities
+/// @{
 
-        /// UI folder
-        string uiFolder;
+/// Set of possible UI properties for an element
+struct UIProperties
+{
+    /// UI name
+    string uiName;
 
-        /// Enumeration
-        StringVec enumeration;
+    /// UI folder
+    string uiFolder;
 
-        /// Enumeration Values
-        vector<ValuePtr> enumerationValues;
+    /// Enumeration
+    StringVec enumeration;
 
-        /// UI minimum value
-        ValuePtr uiMin;
+    /// Enumeration Values
+    vector<ValuePtr> enumerationValues;
 
-        /// UI maximum value
-        ValuePtr uiMax;
+    /// UI minimum value
+    ValuePtr uiMin;
 
-        /// UI advanced element
-        bool uiAdvanced = false;
-    };
+    /// UI maximum value
+    ValuePtr uiMax;
 
-    /// Get the UI properties for a given nodedef element.
-    /// Returns the number of properties found.
-    unsigned int getUIProperties(ConstValueElementPtr nodeDefElement, UIProperties& uiProperties);
+    /// UI advanced element
+    bool uiAdvanced = false;
+};
 
-    /// Get the UI properties for a given element path. If the path is to a node, a target
-    /// identifier can be provided.
-    /// Returns the number of properties found.
-    unsigned int getUIProperties(const string& path, DocumentPtr doc, const string& target, UIProperties& uiProperties);
+/// Get the UI properties for a given nodedef element.
+/// Returns the number of properties found.
+unsigned int getUIProperties(ConstValueElementPtr nodeDefElement, UIProperties& uiProperties);
 
-    /// Interface for holding the UI properties associated shader port
-    struct UIPropertyItem
-    {
-        string label;
-        ValuePtr value;
-        ShaderPort* variable = nullptr;
-        UIProperties ui;
-    };
+/// Get the UI properties for a given element path. If the path is to a node, a target
+/// identifier can be provided.
+/// Returns the number of properties found.
+unsigned int getUIProperties(const string& path, DocumentPtr doc, const string& target, UIProperties& uiProperties);
 
-    /// A grouping of property items by name
-    using UIPropertyGroup = std::multimap<string, UIPropertyItem>;
+/// Interface for holding the UI properties associated shader port
+struct UIPropertyItem
+{
+    string label;
+    ValuePtr value;
+    ShaderPort* variable = nullptr;
+    UIProperties ui;
+};
 
-    /// Utility to group UI properties items based on Element group name from an element.
-    /// Returns a list of named and unnamed groups.
-    void createUIPropertyGroups(ElementPtr uniformElement, DocumentPtr contentDocument, TypedElementPtr materialElement,
-                                const string& pathSeparator, UIPropertyGroup& groups,
-                                UIPropertyGroup& unnamedGroups, ShaderPort* uniform = nullptr);
+/// A grouping of property items by name
+using UIPropertyGroup = std::multimap<string, UIPropertyItem>;
 
-    /// Utility to group UI properties items based on Element group name from a VariableBlock.
-    /// Returns a list of named and unnamed groups.
-    void createUIPropertyGroups(const VariableBlock& block, DocumentPtr contentDocument, TypedElementPtr materialElement,
-                                const string& pathSeparator, UIPropertyGroup& groups,
-                                UIPropertyGroup& unnamedGroups);
+/// Utility to group UI properties items based on Element group name from an element.
+/// Returns a list of named and unnamed groups.
+void createUIPropertyGroups(ElementPtr uniformElement, DocumentPtr contentDocument, TypedElementPtr materialElement,
+                            const string& pathSeparator, UIPropertyGroup& groups,
+                            UIPropertyGroup& unnamedGroups, ShaderPort* uniform = nullptr);
 
-    /// @}
+/// Utility to group UI properties items based on Element group name from a VariableBlock.
+/// Returns a list of named and unnamed groups.
+void createUIPropertyGroups(const VariableBlock& block, DocumentPtr contentDocument, TypedElementPtr materialElement,
+                            const string& pathSeparator, UIPropertyGroup& groups,
+                            UIPropertyGroup& unnamedGroups);
+
+/// @}
 
 } // namespace MaterialX
